Thread Annotation through BindingClassField
Summary:
In order to model class attribute semantics like Final and ClassVar, we need to
be aware of type qualifiers downstream of ClassField calculations.

The first step to making this possible is to preserve the annotation in the
bindings step, which is what this commit does for class body annotations; in
the process we no longer use `AnnotatedType` since that would be redundant (and
in the long run we will not be able to reuse `AnnotatedType`, which is explicitly
dropping the qualifiers).

This commit is not yet handling self attributes assigned in methods, I put
a TODO in and will handle that next since it requires a refactor to the method
binding tables.

We are also not yet making use of the qualifiers, but at this stage the plumbing
to do so is in place, since we have access to the `Annotation` that has
qualifiers inside of `solve_class_field`.

Reviewed By: ndmitchell

Differential Revision: D67303944

fbshipit-source-id: ff0429d19a2912f6331d85d8f9f7ed8ff5a8ffbf