Split call_method into call_method and call_method_or_error
Summary:
I'm working towards adding support for iteration over a class object. In order to reuse our existing iteration code but drop in `call_metaclass_method` as a replacement for `call_method`, I want to create a `call_method` variant that behaves the same way as `call_metaclass_method` - i.e., returns `None` when the attribute isn't found.

I've renamed the original `call_method` to `call_method_or_error` to reflect the fact that it logs an error when the attribute isn't found.

Reviewed By: yangdanny97

Differential Revision: D67615540

fbshipit-source-id: aa289813029a077d00f669d62f178e0c6671a953