Fix crashes cause by missing `ExceptionGroup` class
Summary:
ignore_exception

D67656544 introduced Pyre2 crashes: The diff assumes that `BaseExceptionGroup` always exists. But this assumption doesn’t hold: exception group was introduced in Python3.11, while most project in FBCode is still on Python3.10.

Fortunately I think the fix is simple: Given that exception group is only relevant when the `except*` syntax appears, we just need to re-organize the logic such that `BaseExceptionGroup` is gated behind the same flag that signals the presence of `except*`.

Reviewed By: stroxler, ndmitchell

Differential Revision: D67765293

fbshipit-source-id: e4db7c71ae6b08a809e50c189e62bdbdce28c2c7