Add support for splat args
Summary:
We can splat arguments to function calls. A splat argument consumes some number of positional parameters.

If the splat arg is a fixed-len tuple, the splat arg behaves as if you had passed each element individually as positional arguments.

If the splat arg has indeterminate length, the arg consumes all remaining positional parameters, which can be zero.

Regarding the implementation itself, I think the code is getting fairly complex, but I spent a long time trying different approaches and finally settled on this one:

1. Added a CallArg::Star variant, to avoid accidentally handling splat args for pseudo-calls where they are syntactically invalid.
2. Added "CallArgPreEval" enum, which resolves splat args to either fixed-length or indeterminate forms. An arg can consume 0 or more params.
3. Keep track of the number of provided args by counting how many params are consumed, which required moving the too-many-args error until after all non-keyword args are handled.

I'm not very attached to any of the names I've introduced, so feedback on alternate names is welcome.

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D67657956

fbshipit-source-id: f02012f5df28d6f8f28468d161cc8e4d7e48dd25