Do 'is not' narrowing for more types
Summary:
Allow 'x is not <blah>' to narrow the type of x in more cases where we know that two equal <blah> values are always at the same memory location:
* Boolean literals
* Enum members

Support for enums is partial because I couldn't figure out how to get a list of all enum members; I'll sort that out in another diff.

Also adds a test for `is not <class object>`, which we cannot narrow on due to the covariance of `builtins.type`.

Reviewed By: stroxler

Differential Revision: D67766719

fbshipit-source-id: 2c9b8c366994398b0cbf5255e7b9dac02ff79c29