Remove codenav functionality from protocol.py
Summary:
In this diff, codenav functionality is removed from protocol.py. This is some of the lower-level logic, and I think I've found all the codenav-related functionality, but let me know if it looks like there's anything I've missed.

In this diff, we delete the dataclasses used for marshalling data into and out of the JSON requests and responses used to communicate with the Pyre daemon. These have all been deprecated in prior diffs, so it is safe to delete them, and there are no functional changes.

Reviewed By: ljw1004

Differential Revision: D67775314

fbshipit-source-id: c1b37d0e15343caceeb4289f08f454b11bd3adef