Clean up unnecessary `Arc` in class field plumbing
Summary:
The `Arc<Type>` slipped in here because at one point these things were
direct class field bindings directly sourced from answers. But the logic
is now quite different since we have to pass through various phases of
instantiation as we go from class field to class member to attribute, and
the use of `Arc` is stale.

On trunk, it's making our code more complicated and less efficient; it also
was getting in the way of refactors I need to make.

Reviewed By: ndmitchell

Differential Revision: D67904494

fbshipit-source-id: 8c55e250b61c481525029afd7ab158a2d13e0c29