If an iterator annotation exists, it will be the return type of the function
Summary:
For a function containing yield expressions, if an iterator annotation already exists, we do not construct a generator type. Instead, we can directly return the iterator annotation.

I would like to improve the generator and iterator checks to not use strings, but I'm not sure if there's a way to use some instance check to do so.

For example in

           `self.stdlib
                        .generator(yield_expr_type, Type::any_implicit(), ret)
                        .to_type()`

we are converting the expression to a type already.

If there's a better way to check if such an expression is a generator without looking at the name, I can modify the checks.

Reviewed By: stroxler

Differential Revision: D67814625

fbshipit-source-id: b2e95c7d0209a40f6c3bbd9454c78ea15a0f864e