Remove dead badges from github landing page
Summary:
We have links in the readme that become badges about CI status in the
github landing page. But over the last year we removed the `pyre` and
`lint` signals (because we decided that internal CI should be the source
of truth for these, keeping them perfectly in sync was challenging and
low-value).

We should get rid of the defunct links.

Reviewed By: yangdanny97

Differential Revision: D67910322

fbshipit-source-id: 0de54bc53bf6d3e72f7d8aeabebff124971d3ffc