Patch failing callable instance attr test
Summary:
(See bigger-picture context about the stack below the diff specific notes)

**About This Diff**

The next thing I want to do for class attribute semantics is implement an
end-user-observable behavior change on top of the initialization data that
we've now threaded through bindings. The first thing I'll tackle are
instance-only callable attributes where the callback should be treated as
plain data rather than a bound method.

We already have a test for this, but I want to patch a few things first:
- the name of the test could be better, I updated it
- the test is implicitly coupled to lambda type inference, but that's not
  in scope - a bidirectional type checker can't actually infer precise types
	for lambdas in general. To fix this, I provide an annotation.
- there was an assert saying we ought to be able to propagate literals across
  calls, which I don't think is really true (especially in light of the
	bullet above!). So I'm just using an int for the assert now... that part of the
	test actually already passes, because even if a call fails to type check we
	are generally able to infer the return type.

The next diff will make this entire test case pass.

**About this stack**

I'm threading initialization code through to attribute lookups, which is necessary
to correctly handle instance-only callable attributes vs bound methods (which
is what the test in this diff looks at) and correctly error if we try to access an
instance-only attribute on a class object.

See D67868873 for a longer description with examples.

Reviewed By: rchen152

Differential Revision: D67908365

fbshipit-source-id: ad80e3a8b813f4fbf4db043d5091cc6c0898426e