Read bundled typeshed from Pyre2 binary
Summary:
bundleandsave

Up till now, for Pyre2 to be fully functional one always needs to pass the right path of `typeshed` as a search path to it. By default, some "default" typeshed path was baked in, but there's no guarantee that the path would exist on every machine (in fact the path won't exist on machines that do not have a working copy of FBCode), and there's no guarantee that files under the path would be the same for everyone. In addition, having typeshed separated from the main Pyre2 binary creates a deployment annoyance where we'll have to bundle both the Pyre2 and the typeshed sources together (as was done for Pyre1) when doing any form of release.

This diff attempts to load embeded typeshed into the Pyre2 binary directly, using Rust's `include_bytes!()` functionality. At runtime, we pull the bytes of the tarball from the data section of the executable, decompress all the files into memory, build an index for it, and teach all loaders to look into it before falling back into other search paths.

The tricky part of this diff is: how do we determine what path to use for `include_bytes()!`? As mentioned in D67954102, we build Pyre2 in 2 ways under both Buck and Cargo, and Buck and Cargo unfortunately has different ideas of where they want to put the bundled typeshed archive: In Buck, I could simply add the archive path to `mapped_srcs` of the Rust binary target, and then directly refer to it with relative source paths. In Cargo, since generated files must be located under `$OUT_DIR`, we couldn't really refer to the archive path directly in the source dir, but have to compute the location in the output dir.

What I ended up doing was to have both our Buck build and Cargo build locate the typeshed archive in the same path `${OUT_DIR}/typeshed.tar.zst`. This works out-of-box for Cargo, but for Buck we'll need to "fake" an `OUT_DIR` environment variable in Pyre2's binary rule. It's kind of ugly, but it avoids another can of worm which is to separate our internal and OSS build.

Reviewed By: ndmitchell

Differential Revision: D67955930

fbshipit-source-id: 7a053ab2b91592902583fc037d46f684d2bd60f6