Merge `--output` and `--report-errors` flags in `pyre2 check`
Summary:
duplicatedog

There's the `--output` flag in `pyre2 check` that writes type error output into a file. But there's also the `--report-errors` flag in `pyre2 check` that attempts to do the same thing, except that for `--output` the result is written in JSON format but for `--report-errors` the result is written in plain text.

I think this is a pretty confusing situation, and this diff proposes we clean the flags up as follows:
- First, merge `--output` and `--report-errors` into one single flag. I'll just use `--output` since it's shorter. If the flag is not presented, write to stdout.
- Next, add a `--output-format` flag that can take the value of either `text` and `json` (default is `text`).

The pre-existing `--output` flag can be emulated in the new world via `--output XXX --output-format json`. The pre-existing `--report-errors` flag can be emulated in the new world via `--output XXX` (or `--output XXX --output-format text`).

NOTE: `--output-format json` won't take effect when no `--output` is presented. I would love to make this work eventually (i.e. having it print out json to stdout, which is important to interface with toolings like `pyre-upgrade` because `pyre-upgrade` expects to take its input from stdout in json format). But not changing it for now given that the pre-existing behavior is to not just print out type errors, but also perform additional expectation checks in the absence of the `--output` flag.

Reviewed By: ndmitchell

Differential Revision: D68045160

fbshipit-source-id: 68a5e8ddf3d40c9ffffdc4a929bbda7649c94bb0