Support both flask 2 and flask 3 DSL generator (#960)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/960

Flask 3 moved the route decorators inside the sansio folder which required updating the model. However due to some pyre limitation we are not able to infer the type of aparrapo.route for flask 3 leading to False negative detection which makes the opensource test to fail. While we work on the fixing the false negative keeping the support for both flask 2 and 3.

Reviewed By: arthaud

Differential Revision: D68209629

fbshipit-source-id: 8e1fd6dcfe27a01a1f636cc674c14628998e24b9