Create artificial defines that returns the calls to decorators
Summary:
So that in higher order call graph building, we can redirect the calls to
function `foo` into some artificial define, whose body is `return
decorator(foo)`. This is an optimization, by avoid repeatedly computing the
return values of call expression `decorator(foo)` -- which would now be stored
in the summary of the artificial define.

This diff also adds a new kind `DecoratorDefine` for `Target`s, so that this
new kind can be used to distinguish the artificially created defines from the
defines that come from the source code.  This can be further used when
computing the higher order call graph of the artificial defines, which requires
calling `Target.get_module_and_definition` to find their `Define.t`. We can
look up the `Define.t` in a different way for the artificial defines, which
are identified by the new kind.

Reviewed By: arthaud

Differential Revision: D68024534

fbshipit-source-id: 878828384488ace44cd32156b08978361b7f0d1f