Make pyre able to read configuration from `pyproject.toml` (#799)
Summary:
**Pre-submission checklist**
- [x] I've ran the linters locally and fixed lint errors related to the files I modified in this PR. You can install the linters by running `pip install -r requirements-dev.txt && pre-commit install`
- [x] `pre-commit run`

This PR resolves https://github.com/facebook/pyre-check/issues/695.
I'm not sure we should look for `pyproject.toml` silently when `.pyre_configuration` is not valid, or only look for it when a flag provided.But, I prefer the first resoulution because the second need to provide a flag in all runs.However, I want to hear you opinion.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/799

Test Plan:
Look for if the unittest passed.
As the unittest is not avalible this time(see footnote), I ran test in my local machine with python3.10:
![image](https://github.com/facebook/pyre-check/assets/120731947/fdc82209-5a35-4816-aa28-9c8d1766fba6)

Footnote:
* GitHub Action workflow `Test` fails due to `match` statement is incompatible with Python3.8 and 3.9, and is not related to this PR.To make this easier to review and avoid confliting, I will rebase this PR on the top of main if the `match` statement refactored with `if-elif-else` statement but not do it in this PR.
* As a post-process, we need to delete unneeded `.pyre_configuration` at the root of this project after the nightly version of Pyre with this diff deployed.I intergrated configurations into `pyproject.toml`.

Reviewed By: grievejia

Differential Revision: D68154057

Pulled By: yangdanny97

fbshipit-source-id: 2902549f6ad8aca25f1c67939250615ff2489a3b