support typed dictionary subscript
Summary:
Only allow subscripting with known string literal keys, without allowing attr lookups.

The downside of reusing the attr lookup logic is that the error messages are kind of funky, but we can add a new AttributeBase type and thread some context through the lookup if we want to improve that.

Reviewed By: stroxler

Differential Revision: D68288556

fbshipit-source-id: b3b4d7a9be1343aebe103627822b10d4401e732f