Redirect calling decorated functions to the artificially created defines
Summary:
This diff properly handles the calls to decorated functions / methods. In
particular, consider AST call node `foo(1)` where define `foo` is decorated by
`decorator`.  Then we:
1. (Last diff) Create an artificial define (with name like `foo.$decorator$`)
that returns expression `decorator(foo)`.
2. (This diff) Redirect the call from `foo(1)` to `foo.$decorator$()(1)`. Note
this is *not* `foo.$decorator$(1)` -- we actually need to use the return value
of the artificial define. Such redirection is introduced at the end of building
the original call graphs, so that the higher order call graph building simply
needs to read from the original call graphs.

The benefit is that, we can reuse the summary of `foo.$decorator$` at all call
sites of `foo`. Otherwise, we would need to compute the return values of
`foo.$decorator$` at all call sites of `foo`.

A few noteworthy details:
1. We introduce `decorator_targets` in `CallCallees`, so that
`decorator_targets` stores the calls to the artificially created defines, which
allows setting dependence on them (in order to prompt the global fixpoint to
compute summaries for them).
2. For the artificially created defines, we introduce artificial module names,
so that they can be analyzed in the same way as normal defines.

Reviewed By: arthaud

Differential Revision: D68163501

fbshipit-source-id: 68ec272f41e564fe4932506a951d7536c579a27b