Add plumbing for send type inference
Summary:
Here, we add the keys and bindings needed for send type inference.

The only thing that remains is:

In answers.rs for the   `Binding::SendTypeOfYield(_)`, we must lookup the return annotation and extract the send type from there if it exists.

The key would be                 `&KeyAnnotation::ReturnAnnotation(function_name.clone())`. However, using self.get to retrieve the annotation causes panic if the key is not found.

It looks like we can consider  `std::panic::catch_unwind` but it seems this is unsafe behavior.

Should we modify the `get` method to return an optional?

This is something to address next. Once we do this, the send type will be correctly inferred.

Reviewed By: stroxler

Differential Revision: D68468322

fbshipit-source-id: 4464a1159defed48062d5747ac4e63a237ae3a81