Generalize yields to expressions
Summary:
We must consider `yield` and `yield from` and we must also account for the order.

Consider this example:

` def nested_generator():`
    yield 1
    yield from another_generator()  # E: TODO: YieldFrom(ExprYieldFrom - Answers::expr_infer
    yield 3

`def another_generator():`
    yield 2

The type of `nested_generator()` should be `Generator[Literal[1, 2, 3], Unknown, None]`.

This means that we must consider the order where  `yield from` appears in the `yields` list.

For this reason, we change the type from `ExprYield` to `Expr`.

Next, we will collect `YieldFrom` in the order it appears in the program so that it can appear in the final generator type.

Reviewed By: stroxler

Differential Revision: D68584445

fbshipit-source-id: 249b1e4b266feb0ec8e00d974b04a6b99df7be0c