Use `attr.rs` to handle enum `_value` lookups
Summary:
The normal layer of attribute lookups is that they flow through
`attr.rs`, which forms a pretty nice abstraction boundary between
low-level class code where we are worrying about things like the MRO
and the rest of Pyre2.

One of the few exceptions is the direct use of attribute lookups when
looking up the `_value` type for Enums. There's no actual reason for this
though, we can pretty easily use normal attribute lookups here.

Reviewed By: ndmitchell, rchen152

Differential Revision: D68599172

fbshipit-source-id: 2f1e71b91724d372e18f12bd870ce0faaade7900