Delay expanding yield expressions by adding a new binding for yield values
Summary:
We would like to support `YieldFrom` by generalizing our implementation.

In the binding layer, we are extracting the yield expression too soon, so by the time we are solving the bindings, we cannot tell if an expression is coming from Yield or YieldFrom.

I think we can introduce two bindings to distinguish the two, and then solve them accordngly.

In this diff, I add the first binding and will follow up with a YieldFrom binding.

Reviewed By: stroxler

Differential Revision: D68589490

fbshipit-source-id: c16ece1d653ff944954d6e9e5aa3299dd743ab30