Handle typed dict class object iteration directly
Summary:
The error message for `TypedDict` class object iteration is confusing,
because we're pretending we have a handle on of the `dict` class object.

This is actually not correct: a typed dict *instance* is a `dict` but
a `TypedDict` class object is most certainly not the `dict` class object.

I investigated the behavior of `TypedDict` at runtime and it actually is
iterable as a protocol, but only because the metaclass bakes in that
it is *not* iterable by providing an implementation that always raises:
https://codapi.org/python/

So it does make sense to special case, but we shouldn't be doing this
promotion business for the class object case, and our error message ought to
be clearer.

Reviewed By: rchen152

Differential Revision: D68636615

fbshipit-source-id: e298d8c8f644a50d66c291ab99637334e5aab647