Check for TypedDict directly, not through promote
Summary:
It is risky to rely on `promote` in the way we are doing: it could
potentially produce any `Type`, and we are assuming that *not* getting
a `ClassType` means we must have a `TypedDict` which is true today but
might not remain so.

Let's just check directly. We can also customize the error message a bit;
there's an edge case here (because typed dict's metaclass actually *does*
define iteration, but we know that it always raises a runtime error) so
it's good to provide some evidence of our reasoning.

Reviewed By: rchen152

Differential Revision: D68636617

fbshipit-source-id: 3258fa01a074843a174d3772ef9eb276e9782e4b